pyproject.toml
[tool.black]
line-length = 88
target-version = ['py38']
exclude = '''
/(
    | migrations
)/
'''
[tool.poetry]
name = "museum_backend"
version = "0.1.0"
description = "Backend for virtual museum"
authors = ["database <karkiaabishkar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.3"
djangorestframework = "3.12.4"
dj-database-url = "0.5.0"
psycopg2-binary = "2.8.6"
asgiref = "3.3.4"
gunicorn = "20.1.0"
importlib-metadata = "^4.2.0"
typing-extensions = "^3.10.0"
sentry-sdk = "^1.3.1"
django-watchman = "^1.2.0"
drf-spectacular = "^0.17.3"
django-filter = "^2.4.0"
djangorestframework-simplejwt = "^4.8.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.4"
coverage = "5.5"
pytest-cov = "^2.12.0"
pytest-django = "^4.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov --cov-append --cov-report term-missing"
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = "tests.py test_*.py *_tests.py"